refact: BallNumber 값 -> record 타입으로 박싱
