+++ src/main/java/baseball/util/NumberPicker.java
@@ -1,0 +1,6 @@
+package baseball.util;
+
+public interface NumberPicker {
+
+  int pick();
+}
